Run graphs (#72) & Icons (#75)
Run graphs added for total consumption and installation's consumption,
as well as validation so that if returned JSON contains too many data,
plot is not loaded, so that browser doesn't crash.
Also changed JSONtoReturn.java and MongoDBQueries.java to return
numberOfDays ( in order to calculate toTick value) and
defaultAggregationUnit, meaning that even if the user specifies an
aggregation unit, I needed this one returned for calculations.