issue #147: fixes
Regarding this issue (#147) the following came up:
1. If user enters whatever credentials then cames up a
NullPointerException in Java 'cause user doesn't exist in db.
(authenticate(Utils.java:118))
2 If user is successfully authenticated but has no projects it isn't
possible to create new ones with the current implementation. That's
because the usr_id has never been sent to client.
My suggestion on 2, is that projects' schema doesn't need to contain
usr_id because the usr_id is never needed on the client side.
However if it suits you to better to include usr_id in the schema, then
upon btn click (login), the usr_id should be returned and the GET
projects request will happen after that.