fixes #302
Best thing I could do was to move the legend to it's current position
and add a transparent background, since the legend is part of the svg
graph element so styling it can be tricky. (e.g. moving it to the bottom
of the graph with a vertical alignment of runs cannot be implemented
properly).
About the displayed names now they are in correspondence with the
waveforms.