update #409
a) in the installation form we would like to have the operating hours of
the installation, with a control like in TOU pricing. Perhaps per
quarter of the hour.

b) In the appliance add a new tickbox with the filed Lighting. If check
a text area should appear where the user will import an array of comma
separated values. If unchecked the text area should disappear and get
emptied. Name the text area Montlhy Cosumptions.